requirements.txt
absl-py==2.3.1
agents==1.4.0
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
attrs==25.3.0
av==12.3.0
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cryptography==45.0.5
deprecation==2.1.0
dnspython==2.6.1
exceptiongroup==1.3.0
filelock==3.16.1
firebase-admin==6.9.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.9.2
grpcio==1.70.0
grpcio-status==1.70.0
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.13.1
libclang==18.1.1
livekit==0.4.6
llvmlite==0.41.1
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.42.1
networkx==3.1
notion-client==3.0.0
numba==0.58.1
numpy==1.24.3
oauthlib==3.3.1
openai-whisper==20250625
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
postgrest==0.16.11
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.9.0
pylibsrtp==0.10.0
pyOpenSSL==25.1.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
realtime==1.0.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.8
sip==6.8.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.7.7
streamlit==1.40.1
streamlit-webrtc==0.48.2
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
termcolor==2.4.0
tiktoken==0.7.0
toml==0.10.2
tomli==2.2.1
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
types-protobuf==5.29.1.20241207
typing_extensions==4.13.2
tzdata==2025.2
UNKNOWN @ file:///Users/sricharan/Desktop/hackathon/agents
uritemplate==4.1.1
urllib3==2.2.3
websockets==12.0
Werkzeug==3.0.6
whisper==1.1.10
wrapt==1.17.2
zipp==3.20.2
